refactor turn files to match starter repo layout
